Migrate BuiltIn Functions and Methods to use the FunctionAndMethodAPI
